+++ Software Project/1 - Backend - Spring - CleanArchitecture/2 - Application/src/main/java/de/dhbw/ase/seriephant/episode/EpisodeApplicationService.java
@@ -1,5 +1,7 @@
 package de.dhbw.ase.seriephant.episode;
 
+import de.dhbw.ase.seriephant.domain.actor.Actor;
+import de.dhbw.ase.seriephant.domain.actor.ActorRepository;
 import de.dhbw.ase.seriephant.domain.episode.Episode;
 import de.dhbw.ase.seriephant.domain.episode.EpisodeRepository;
 import de.dhbw.ase.seriephant.domain.season.SeasonRepository;
@@ -8,17 +10,20 @@ import org.springframework.stereotype.Service;
 
 import javax.xml.bind.ValidationException;
 import java.time.LocalDate;
+import java.util.ArrayList;
 import java.util.List;
 
 @Service
 public class EpisodeApplicationService {
     private final EpisodeRepository episodeRepository;
     private final SeasonRepository seasonRepository;
+    private final ActorRepository actorRepository;
 
     @Autowired
-    private EpisodeApplicationService(EpisodeRepository episodeRepository, SeasonRepository seasonRepository) {
+    private EpisodeApplicationService(EpisodeRepository episodeRepository, SeasonRepository seasonRepository, ActorRepository actorRepository) {
         this.episodeRepository = episodeRepository;
         this.seasonRepository = seasonRepository;
+        this.actorRepository = actorRepository;
     }
 
     /************************************************************************************************************************************/
@@ -32,7 +37,23 @@ public class EpisodeApplicationService {
         \_____|_|  \___|\__,_|\__\___|
      */
     public Episode saveEpisode(Episode episode) throws ValidationException {
-        return this.episodeRepository.save(episode);
+        if (episode != null && (episode.getId() == null || !this.episodeRepository.existsById(episode.getId()))) {
+            List<Actor> actors = new ArrayList<>();
+            if (episode.getActors() != null) {
+                episode.getActors().forEach(actor -> {
+                    if (this.actorRepository.existsById(actor.getId())) {
+                        actors.add(actor);
+                    }
+                });
+            }
+            episode.setActors(actors);
+            if (episode.getSeason() != null && episode.getSeason().getId() != null && this.seasonRepository.existsById(episode.getSeason().getId())) {
+                episode.setSeason(this.seasonRepository.getById(episode.getSeason().getId()));
+                return this.episodeRepository.save(episode);
+            }
+            throw new ValidationException("Season is not valid.");
+        }
+        throw new ValidationException("Episode is not valid.");
     }
 
     public Episode saveEpisode(String title, LocalDate releaseDate, Integer episodeNumber, Long seasonId) throws ValidationException {
@@ -75,15 +96,28 @@ public class EpisodeApplicationService {
               |_|
     */
     public Episode updateEpisode(Episode episode) throws ValidationException {
-        if (this.episodeRepository.existsById(episode.getId())) {
+        if (episode != null && episode.getId() == null && this.episodeRepository.existsById(episode.getId())) {
             Episode foundEpisode = this.episodeRepository.getById(episode.getId());
             foundEpisode.setTitle(episode.getTitle());
             foundEpisode.setReleaseDate(episode.getReleaseDate());
             foundEpisode.setEpisodeNumber(episode.getEpisodeNumber());
-            foundEpisode.setSeason(episode.getSeason());
-            return this.episodeRepository.save(foundEpisode);
+            List<Actor> actors = new ArrayList<>();
+            if (episode.getActors() != null) {
+                episode.getActors().forEach(actor -> {
+                    if (this.actorRepository.existsById(actor.getId())) {
+                        actors.add(actor);
+                    }
+                });
+                foundEpisode.setActors(actors);
+            }
+            if (episode.getSeason() != null && episode.getSeason().getId() != null && this.seasonRepository.existsById(episode.getSeason().getId())) {
+                foundEpisode.setSeason(this.seasonRepository.getById(episode.getSeason().getId()));
+                return this.episodeRepository.save(foundEpisode);
+            }
+            throw new ValidationException("Season is not valid.");
         }
-        throw new ValidationException("Id of Episode is not known.");
+        throw new ValidationException("Episode is not valid.");
+
     }
 
     /************************************************************************************************************************************/
@@ -96,7 +130,9 @@ public class EpisodeApplicationService {
         |_____/ \___|_|\___|\__\___|
     */
     public void deleteEpisode(Long episodeId) {
-        this.episodeRepository.deleteById(episodeId);
+        if (this.episodeRepository.existsById(episodeId)) {
+            this.episodeRepository.deleteById(episodeId);
+        }
     }
     /************************************************************************************************************************************/
 }
